+++ server/resources/grade_distributions.py
@@ -33,37 +33,42 @@ class AddGradeDistribution(Resource):
     @jwt_required
     def post(self):
         args = parser.parse_args()
-        grade_distribution = GradeDistribution(
-            lecturer_id=args['lecturer_id'],
-            user_id=get_jwt_identity()['id'],
-            term_id=args['term_id'],
-            course_id=args['course_id'],
-            course_code=args['course_code'],
-            english=args['english'],
-        )
-
+        user_id = get_jwt_identity()['id']
         file = args['image']
 
         if file:
-            filename = grade_distribution.image = grade_distribution.generateImageName()
-            imgdata = base64.b64decode(file)
+            grade_distribution = GradeDistribution()
+            filename = grade_distribution.generateImageName()
+            grade_distribution.create({
+                'lecturer_id': args['lecturer_id'],
+                'user_id': user_id,
+                'term_id': args['term_id'],
+                'course_id': args['course_id'],
+                'course_code': args['course_code'],
+                'english': args['english'],
+                'image': filename
+            })
+
+            imgdata = base64.b64decode(file.split(';base64,', 1)[1])
 
             with open(server.config['UPLOAD_FOLDER'] + '/' + filename, 'wb') as f:
                 f.write(imgdata)
 
             file_size = os.stat(os.path.join(server.config['UPLOAD_FOLDER'], filename)).st_size
-            print('393343', file=sys.stderr)
+
             if file_size > 5000000:
                 os.remove(os.path.join(server.config['UPLOAD_FOLDER'], filename))
                 return response({
                     'errors': [
                         'File size must be lower than 5 MB'
                     ]
-                })
+                }, 400)
 
-            if grade_distribution.save() is True:
+            if grade_distribution.validate() is True:
+                grade_distribution.save()
+
                 return response({
-                    'message': 'Grade distribution saved successfully'
+                    'grade_distribution': grade_distribution.data()
                 })
 
             return response({
@@ -80,14 +85,20 @@ class AddGradeDistribution(Resource):
 
 class DeleteGradeDistribution(Resource):
     @jwt_required
-    def post(self):
-        args = parser.parse_args()
-        grade_distribution = GradeDistribution(
-            _id=args['id'],
-            user_id=get_jwt_identity()['id'],
-        )
-        grade_distribution.delete()
+    def post(self, dist_id):
+        #args = parser.parse_args()
+        user_id = get_jwt_identity()['id']
+        grade_distribution = GradeDistribution().where([['id', '=', dist_id],
+                                            ['user_id', '=', user_id]]).first()
+
+        if grade_distribution.exists() is True:
+            grade_distribution.delete()
+            return response({
+                'message': 'Grade Distribution has deleted successfully'
+            }, 202)
 
         return response({
-            'message': 'Grade Distribution has deleted successfully'
-        })
+            'errors': [
+                'Grade distribution could not found'
+            ]
+        }, 404)
